Build using Docker
Pin Go version to 1.16.5, latest at time of development.
Use a multi-stage build for security and slim final image size.
Use Alpine for final image base, pin to latest stable.
Create an 'api' user to run the application as a non-root user.

Add a .dockerignore to omit the README and scripts/ from the build.