Parse POST request body
I normally wouldn't make a commit at this point, as the handler isn't
finished, but I wanted to demonstrate a point at which I would run the
scripts in `scripts/` to begin testing this handler. At this point, I'm
expecting `post_request` to print out the request body as given, and
`bad_data` to log a JSON parsing error.