Screener stepped form functionality
Achieve sequentially-stepped form by creating custom state object
`form`. This is initialized in App.js using the data fetched from the
backend API service, and updated in DiagnosticScreener.js though an
update function passed as a prop.

DiagnosticScreener.js uses a simple stepper to keep track of the current
question, sending a POST request with the responses after the final
question has been answered. This relies upon the questions requiring a
single response and being sequential. This also does not allow for
viewing or changing previous responses.

If we need the user to be able to move backwards through the questions,
then we could update the `next` function to accept a `target` argument.

If the logic for questions or forms increases in complexity beyond what
can reasonably be managed in this state/update fashion, I would look to
a library like Redux.js to manage the state in a more robust fashion.